add tanglism api in web app

+++ tanglism-morph/src/shape.rs
@@ -1,16 +1,16 @@
 use serde_derive::*;
 use chrono::NaiveDateTime;
-
+use bigdecimal::BigDecimal;
 
 /// K线
 ///
 /// 缠论的基础概念
 /// 在缠论中，K线的开盘价和收盘价被忽略，仅包含时刻，最高点，最低点
-#[derive(Debug, Clone, Copy, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct K {
     pub ts: NaiveDateTime,
-    pub low: f64,
-    pub high: f64,
+    pub low: BigDecimal,
+    pub high: BigDecimal,
 }
 
 /// 合并K线
@@ -20,13 +20,13 @@ pub struct K {
 /// 低价低时，满足包含原则，两K线可视为1条K线。在上升时，取两高点的高点为新K线高点，
 /// 取两低点的高点为新K线低点。在下降时，取两高点的低点为新K线高点，取两低点的低点
 /// 为新K线的低点。
-#[derive(Debug, Clone, Copy, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct CK {
     pub start_ts: NaiveDateTime,
     pub end_ts: NaiveDateTime,
     pub extremum_ts: NaiveDateTime,
-    pub low: f64,
-    pub high: f64,
+    pub low: BigDecimal,
+    pub high: BigDecimal,
     pub n: i32,
 }
 
@@ -40,7 +40,7 @@ pub struct CK {
 /// 分型实际可由多于3根K线构成，只要两侧的K线满足包含原则。
 /// 按照缠论的严格定义，分型仅适用与最小级别的K线图，即1分钟K线图上，后续分析都由
 /// 1分钟K线图向上递归构成更大的形态。
-#[derive(Debug, Clone, Copy, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Parting {
     // 分型起始时刻，已考虑K线包含关系
     pub start_ts: NaiveDateTime,
@@ -49,7 +49,7 @@ pub struct Parting {
     // 分型转折时刻
     pub extremum_ts: NaiveDateTime,
     // 转折点价格
-    pub extremum_price: f64,
+    pub extremum_price: BigDecimal,
     // 组成分型的K线数
     pub n: i32,
     // 是否顶分型，非顶即底分型
@@ -61,7 +61,7 @@ pub struct Parting {
 /// 缠论的基础概念
 /// 由相邻的顶分型与底分型构成，不可同底或同顶，同时需满足两分型间有至少1根独立K线，
 /// 即存在1条K线，不属于两侧的分型，且不能因为包含原则属于两侧的分型。
-#[derive(Debug, Clone, Copy, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Stroke {
     pub start_pt: Parting,
     pub end_pt: Parting,
@@ -71,7 +71,7 @@ pub struct Stroke {
 /// 
 /// 在特征序列相邻笔出现包含关系时，合并为一笔
 /// 此时笔并不具有方向性
-#[derive(Debug, Clone, Copy, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct CStroke {
     pub high_pt: Parting,
     pub low_pt: Parting,
@@ -89,7 +89,7 @@ pub struct CStroke {
 /// 顶分型的顶即向上线段的结束。
 /// 底分型的底即向下线段的结束。
 /// 当确定线段终点后，该终点后的笔不再归属于该线段。
-#[derive(Debug, Clone, Copy, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Segment {
     pub start_pt: Parting,
     pub end_pt: Parting,
@@ -106,15 +106,15 @@ pub struct Center {
     // 起始时刻
     pub start_ts: NaiveDateTime,
     // 起始价格
-    pub start_price: f64,
+    pub start_price: BigDecimal,
     // 结束时刻
     pub end_ts: NaiveDateTime,
     // 结束价格
-    pub end_price: f64,
+    pub end_price: BigDecimal,
     // 共享最低点，即所有次级别走势类型的最低点中的最高点
-    pub shared_low: f64,
+    pub shared_low: BigDecimal,
     // 共享最高点，即所有次级别走势类型的最高点中的最低点
-    pub shared_high: f64,
+    pub shared_high: BigDecimal,
     // 是否起始于向上走势中
     pub upward: bool,
     // 级别
@@ -137,11 +137,11 @@ pub struct Trend {
     // 起始时刻
     pub start_ts: NaiveDateTime,
     // 起始价格
-    pub start_price: f64,
+    pub start_price: BigDecimal,
     // 结束时刻
     pub end_ts: NaiveDateTime,
     // 结束价格
-    pub end_price: f64,
+    pub end_price: BigDecimal,
     // 中枢列表
     pub centers: Vec<Center>,
     // 级别
